Move bitcoin address derivation to the bitcoin package
This has a few main benefits. Since this code isn't specific to bitcoin recovery, and more generically applicable to working with bitcoin, we can move the derivation code to the bitcoin package.

The primary use case is that we want our bitcoin config validation code to be able to check to see if the supplied btc address is valid, which depends on running address derivation. Moving the derivation code to the bitcoin package removes a circular dependency.
